+++ app/src/main/java/de/baumann/browser/Browser_1.java
@@ -44,7 +44,6 @@ import android.preference.PreferenceManager;
 import android.provider.MediaStore;
 import android.support.design.widget.Snackbar;
 import android.support.v4.content.ContextCompat;
-import android.support.v4.widget.SwipeRefreshLayout;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.support.v7.widget.CardView;
@@ -112,6 +111,7 @@ public class Browser_1 extends AppCompatActivity implements ObservableScrollView
     private EditText editText;
     private RelativeLayout relativeLayout;
     private HorizontalScrollView scrollTabs;
+    private Toolbar toolbar;
 
 
     // Strings
@@ -192,7 +192,7 @@ public class Browser_1 extends AppCompatActivity implements ObservableScrollView
 
         // find Views
 
-        Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
+        toolbar = (Toolbar) findViewById(R.id.toolbar);
         mWebView = (ObservableWebView) findViewById(R.id.webView);
         relativeLayout = (RelativeLayout) findViewById(R.id.bottom_layout);
         customViewContainer = (FrameLayout) findViewById(R.id.customViewContainer);
@@ -211,21 +211,11 @@ public class Browser_1 extends AppCompatActivity implements ObservableScrollView
         toolbar.setBackgroundColor(ContextCompat.getColor(activity, R.color.colorPrimary_1));
         setSupportActionBar(toolbar);
 
-        SwipeRefreshLayout swipeView = (SwipeRefreshLayout) findViewById(R.id.swipe);
-        assert swipeView != null;
-        swipeView.setColorSchemeResources(R.color.colorPrimary_1, R.color.colorAccent);
-        swipeView.setOnRefreshListener(new SwipeRefreshLayout.OnRefreshListener() {
-            @Override
-            public void onRefresh() {
-                mWebView.reload();
-            }
-        });
-
 
         // setup WebView
 
         helper_webView.webView_Settings(activity, mWebView);
-        helper_webView.webView_WebViewClient(activity, swipeView, mWebView, urlBar);
+        helper_webView.webView_WebViewClient(activity, mWebView, urlBar);
 
         mWebChromeClient = new myWebChromeClient();
         mWebView.setWebChromeClient(mWebChromeClient);
@@ -572,7 +562,7 @@ public class Browser_1 extends AppCompatActivity implements ObservableScrollView
 
     @Override
     public void onUpOrCancelMotionEvent(ScrollState scrollState) {
-        helper_browser.scroll(activity, scrollState, relativeLayout, imageButton, imageButton_left, imageButton_right,
+        helper_browser.scroll(activity, scrollState, relativeLayout, toolbar, imageButton, imageButton_left, imageButton_right,
                 urlBar, mWebView, scrollTabs);
     }
 
@@ -670,6 +660,12 @@ public class Browser_1 extends AppCompatActivity implements ObservableScrollView
                             });
                     snackbar.show();
                     mWebView.loadUrl(URL.replace("openLogin", ""));
+                } else if (URL.contains("exit")) {
+                    sharedPref.edit().putString("openURL", "").apply();
+                    sharedPref.edit().putString("tab_1", "").apply();
+                    File tab_1 = new File(activity.getFilesDir() + "/tab_1.jpg");
+                    tab_1.delete();
+                    finish();
                 } else {
                     mWebView.loadUrl(URL);
                 }
@@ -850,6 +846,10 @@ public class Browser_1 extends AppCompatActivity implements ObservableScrollView
             editText.setVisibility(View.GONE);
         }
 
+        if (id == R.id.action_reload) {
+            mWebView.reload();
+        }
+
         return super.onOptionsItemSelected(item);
     }
 
@@ -908,15 +908,6 @@ public class Browser_1 extends AppCompatActivity implements ObservableScrollView
             } catch (Exception e) {
                 // Error occurred while creating the File
                 Log.e(TAG, "Browser Error", e);
-            }
-
-            if (progress <= 10) {
-                imageButton.setVisibility(View.INVISIBLE);
-                relativeLayout.animate().translationY(0);
-
-                if (scrollTabs.getVisibility() == View.VISIBLE) {
-                    scrollTabs.setVisibility(View.GONE);
-                }
             }
 
             if (progress == 100) {