+++ app/src/main/java/de/baumann/browser/lists/List_readLater.java
@@ -70,6 +70,9 @@ public class List_readLater extends AppCompatActivity {
     private SimpleCursorAdapter adapter;
     private SharedPreferences sharedPref;
 
+    private int top;
+    private int index;
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -133,6 +136,12 @@ public class List_readLater extends AppCompatActivity {
         }
     }
 
+    private void isEdited () {
+        index = listView.getFirstVisiblePosition();
+        View v = listView.getChildAt(0);
+        top = (v == null) ? 0 : (v.getTop() - listView.getPaddingTop());
+    }
+
     private void setReadLaterList() {
 
         //display data
@@ -169,6 +178,7 @@ public class List_readLater extends AppCompatActivity {
         });
 
         listView.setAdapter(adapter);
+        listView.setSelectionFromTop(index, top);
         //onClick function
         listView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
             @Override
@@ -183,6 +193,8 @@ public class List_readLater extends AppCompatActivity {
 
         listView.setOnItemLongClickListener(new AdapterView.OnItemLongClickListener() {
             public boolean onItemLongClick(AdapterView<?> parent, View view, int position, long id) {
+
+                isEdited();
 
                 Cursor row2 = (Cursor) listView.getItemAtPosition(position);
                 final String _id = row2.getString(row2.getColumnIndexOrThrow("_id"));
@@ -312,11 +324,6 @@ public class List_readLater extends AppCompatActivity {
                 return true;
             }
         });
-
-        listView.post(new Runnable(){
-            public void run() {
-                listView.setSelection(listView.getCount() - 1);
-            }});
     }
 
     private void setTitle () {
@@ -338,6 +345,8 @@ public class List_readLater extends AppCompatActivity {
     @Override
     public boolean onPrepareOptionsMenu(Menu menu) {
         super.onPrepareOptionsMenu(menu);
+
+        menu.findItem(R.id.action_favorite).setVisible(false);
 
         if (sharedPref.getInt("keyboard", 0) == 0) {
             // normal